rename gen.lua as roomgen.lua
