+++ lib/roomgen.lua
